remove unnecessary conftest
