+++ services/storage/tests/conftest.py
@@ -11,6 +11,7 @@ import logging
 import sys
 import urllib.parse
 from collections.abc import AsyncIterator, Awaitable, Callable
+from contextlib import AbstractAsyncContextManager, asynccontextmanager
 from pathlib import Path
 from typing import cast
 
@@ -49,6 +50,7 @@ from servicelib.aiohttp import status
 from simcore_postgres_database.storage_models import file_meta_data, projects, users
 from simcore_service_storage.application import create
 from simcore_service_storage.dsm import get_dsm_provider
+from simcore_service_storage.handlers_files import UPLOAD_TASKS_KEY
 from simcore_service_storage.models import S3BucketName
 from simcore_service_storage.s3 import get_s3_client
 from simcore_service_storage.settings import Settings
@@ -58,6 +60,7 @@ from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 from tests.helpers.utils_file_meta_data import assert_file_meta_data_in_db
+from types_aiobotocore_s3 import S3Client
 from yarl import URL
 
 pytest_plugins = [
@@ -474,3 +477,184 @@ def create_simcore_file_id(
         return SimcoreS3FileID(f"{clean_path}")
 
     return _creator
+
+
+@pytest.fixture
+async def with_versioning_enabled(
+    s3_client: S3Client,
+    storage_s3_bucket: S3BucketName,
+) -> None:
+    await s3_client.put_bucket_versioning(
+        Bucket=storage_s3_bucket,
+        VersioningConfiguration={"MFADelete": "Disabled", "Status": "Enabled"},
+    )
+
+
+@pytest.fixture
+async def create_empty_directory(
+    create_simcore_file_id: Callable[[ProjectID, NodeID, str], SimcoreS3FileID],
+    create_upload_file_link_v2: Callable[..., Awaitable[FileUploadSchema]],
+    client: TestClient,
+    project_id: ProjectID,
+    node_id: NodeID,
+) -> Callable[..., Awaitable[FileUploadSchema]]:
+    async def _directory_creator(dir_name: str):
+        # creating an empty directory goes through the same procedure as uploading a multipart file
+        # done by using 3 calls:
+        # 1. create the link as a directory
+        # 2. call complete_upload link
+        # 3. call file_upload_complete_response until it replies OK
+
+        directory_file_id = create_simcore_file_id(project_id, node_id, dir_name)
+        directory_file_upload: FileUploadSchema = await create_upload_file_link_v2(
+            directory_file_id, link_type="s3", is_directory="true", file_size=-1
+        )
+        # always returns a v2 link when dealing with directories
+        assert isinstance(directory_file_upload, FileUploadSchema)
+        assert len(directory_file_upload.urls) == 1
+
+        # complete the upload
+        complete_url = URL(directory_file_upload.links.complete_upload).relative()
+        response = await client.post(
+            f"{complete_url}",
+            json=jsonable_encoder(FileUploadCompletionBody(parts=[])),
+        )
+        response.raise_for_status()
+        data, error = await assert_status(response, status.HTTP_202_ACCEPTED)
+        assert not error
+        assert data
+        file_upload_complete_response = FileUploadCompleteResponse.parse_obj(data)
+        state_url = URL(file_upload_complete_response.links.state).relative()
+
+        # check that it finished updating
+        assert client.app
+        client.app[UPLOAD_TASKS_KEY].clear()
+        # now check for the completion
+        async for attempt in AsyncRetrying(
+            reraise=True,
+            wait=wait_fixed(1),
+            stop=stop_after_delay(60),
+            retry=retry_if_exception_type(AssertionError),
+        ):
+            with attempt, log_context(
+                logging.INFO,
+                f"waiting for upload completion {state_url=}, {attempt.retry_state.attempt_number}",
+            ) as ctx:
+                response = await client.post(f"{state_url}")
+                data, error = await assert_status(response, status.HTTP_200_OK)
+                assert not error
+                assert data
+                future = FileUploadCompleteFutureResponse.parse_obj(data)
+                assert future.state == FileUploadCompleteState.OK
+                assert future.e_tag is None
+                ctx.logger.info(
+                    "%s",
+                    f"--> done waiting, data is completely uploaded [{attempt.retry_state.retry_object.statistics}]",
+                )
+
+        return directory_file_upload
+
+    return _directory_creator
+
+
+@pytest.fixture
+async def populate_directory(
+    create_file_of_size: Callable[[ByteSize, str | None], Path],
+    storage_s3_client: SimcoreS3API,
+    storage_s3_bucket: S3BucketName,
+    project_id: ProjectID,
+    node_id: NodeID,
+) -> Callable[..., Awaitable[None]]:
+    async def _create_content(
+        file_size_in_dir: ByteSize,
+        dir_name: str,
+        subdir_count: int = 4,
+        file_count: int = 5,
+    ) -> None:
+        file = create_file_of_size(file_size_in_dir, "some_file")
+
+        async def _create_file(s: int, f: int):
+            file_name = f"{dir_name}/sub-dir-{s}/file-{f}"
+            clean_path = Path(f"{project_id}/{node_id}/{file_name}")
+            await storage_s3_client.upload_file(
+                bucket=storage_s3_bucket,
+                file=file,
+                object_key=SimcoreS3FileID(f"{clean_path}"),
+                bytes_transfered_cb=None,
+            )
+
+        tasks = [
+            _create_file(s, f) for f in range(file_count) for s in range(subdir_count)
+        ]
+
+        await asyncio.gather(*tasks)
+
+        file.unlink()
+
+    return _create_content
+
+
+@pytest.fixture
+async def delete_directory(
+    client: TestClient,
+    storage_s3_client: SimcoreS3API,
+    storage_s3_bucket: S3BucketName,
+    user_id: UserID,
+    location_id: LocationID,
+) -> Callable[..., Awaitable[None]]:
+    async def _dir_remover(directory_file_upload: FileUploadSchema) -> None:
+        assert directory_file_upload.urls[0].path
+        directory_file_id = directory_file_upload.urls[0].path.strip("/")
+        assert client.app
+        delete_url = (
+            client.app.router["delete_file"]
+            .url_for(
+                location_id=f"{location_id}",
+                file_id=urllib.parse.quote(directory_file_id, safe=""),
+            )
+            .with_query(user_id=user_id)
+        )
+        response = await client.delete(f"{delete_url}")
+        await assert_status(response, status.HTTP_204_NO_CONTENT)
+
+        # NOTE: ensures no more files are left in the directory,
+        # even if one file is left this will detect it
+        list_files_metadata_url = (
+            client.app.router["get_files_metadata"]
+            .url_for(location_id=f"{location_id}")
+            .with_query(user_id=user_id, uuid_filter=directory_file_id)
+        )
+        response = await client.get(f"{list_files_metadata_url}")
+        data, error = await assert_status(response, status.HTTP_200_OK)
+        assert error is None
+        assert data == []
+
+    return _dir_remover
+
+
+@pytest.fixture
+async def create_directory_with_files(
+    create_empty_directory: Callable[..., Awaitable[FileUploadSchema]],
+    populate_directory: Callable[..., Awaitable[None]],
+    delete_directory: Callable[..., Awaitable[None]],
+) -> Callable[..., AbstractAsyncContextManager[FileUploadSchema]]:
+    @asynccontextmanager
+    async def _create_context(
+        dir_name: str, file_size_in_dir: ByteSize, subdir_count: int, file_count: int
+    ) -> AsyncIterator[FileUploadSchema]:
+        directory_file_upload: FileUploadSchema = await create_empty_directory(
+            dir_name=dir_name
+        )
+
+        await populate_directory(
+            file_size_in_dir=file_size_in_dir,
+            dir_name=dir_name,
+            subdir_count=subdir_count,
+            file_count=file_count,
+        )
+
+        yield directory_file_upload
+
+        await delete_directory(directory_file_upload=directory_file_upload)
+
+    return _create_context
